+++ feature/setting/presentation/src/main/kotlin/com/afternote/feature/setting/presentation/component/SettingProfile.kt
@@ -1,0 +1,2 @@
+package com.afternote.feature.setting.presentation.component
+
